tratamento de exceções

+++ src/main/java/com/devsuperior/ics/dscommerce/services/ProductService.java
@@ -1,18 +1,20 @@
 package com.devsuperior.ics.dscommerce.services;
 
-import java.util.List;
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import com.devsuperior.ics.dscommerce.services.exceptions.ResourceNotFoundException;
 
 import com.devsuperior.ics.dscommerce.dto.ProductDTO;
 import com.devsuperior.ics.dscommerce.entities.Product;
+import com.devsuperior.ics.dscommerce.repositories.DatabaseException;
 import com.devsuperior.ics.dscommerce.repositories.ProductRepository;
+import com.devsuperior.ics.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -52,18 +54,32 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		
-	Product entity = repository.getReferenceById(id);
-	copyDtoEntity(dto, entity);
-	entity = repository.save(entity);
-	return new ProductDTO(entity);
+	 try {
+		Product entity = repository.getReferenceById(id);
+	    copyDtoEntity(dto, entity);
+	    entity = repository.save(entity);
+	    return new ProductDTO(entity);
+	    } 
+	 catch (EntityNotFoundException e) {
+		 throw new ResourceNotFoundException("Recurso nao encontrado");
+	  }
 	}
 
 	
 	
-	@Transactional()
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		 repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso nao Encontrado");
+		}
+		try {
+			 repository.deleteById(id);
+		}
+		
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
+		
 	}
